port 03-production-names.t
I thought for a second that it wouldn't be possible to port the tests
that assert parse failure, but then realized that tree-sitter outputs
errors as part of the parse structure! This allows us to *improve* the
tests.

One note, I'm not super happy with the debug output in the error cases.
Here's the one for the `semicolon` test case:

    lex_internal state:1, row:2, column:4
      consume character:'a'
    lexed_lookahead sym:prod_name, size:1
    shift state:3
    process version:0, version_count:1, state:3, row:2, col:5
    lex_internal state:0, row:2, column:5
      consume character:';'
    lex_internal state:0, row:2, column:5
      consume character:';'
    skip_unrecognized_character
    lex_internal state:0, row:2, column:6
      consume character:'a'
    lex_internal state:0, row:2, column:7
      skip character:10
      skip character:' '
      skip character:' '
      consume character:'('
    lexed_lookahead sym:ERROR, size:2
    detect_error

The reason that `;` is matched is because a comment token can begin with
it, but notice that the `comment` symbol is not mentioned anywhere in
the debug output!