Port 06-conditions.t
Make the new tests pass by porting over the remaining LHS grammar rules.

We've matched the behavior of the Perl parser, which matches the
behavior of the CSoar parser. This means that some structures parse
but later crash Soar; I don't honestly think this should count, and will
be revising the tests to prevent Soar-crashing structures from being
parseable.