Fix attr_value_make, which is wrong in the manual
The attributes in attribute-value makes in RHS actions are composed of
RHS values joined by dots; the manual says that they can only be
variables or constants, which is incorrect. The grammar given in the CPP
code is also missing any mention of the dot sections.

Update the grammar to match the actual parser. Update the test cases to
match the new outputs. Add a new test case for other types of RHS values
in attr-value makes.