Add entire CSoar test suite to our test corpus
Note that the productions using `""` instead of `{}` still do not parse.
I can't get CSoar to parse these either, so there's something I'm
missing. Also, none of the commands that aren't `sp` also fail to parse.

Everything else was able to be parsed. Parsing this data led to the
fixes in two earlier commits: the misplaced `+` in valueTest and the
overly strict `attr_value_make` structure.

Note that there is an issue with tree-sitter that causes some of the
output s-expressions to be invalid, resulting in `(UNEXPECTED ' ')` and
`(UNEXPECTED ')')` being output as `(UNEXPECTED ')`. The test
comparisons fail as a result as well, leading to 12 or so test failures.
As we will be updating the parser to handle these sections, I won't
bother patching tree-sitter or anything drastic like that. But filing
a ticket is a TODO.