flatten and add field names to RHS rules
Remove extra tree nodes by eliminating some rules and hiding others
using `_`. Add field names, as well, to make the output as intuitive as
possible. This is following the advice given in
https://tree-sitter.github.io/tree-sitter/creating-parsers#
in the the "Writing the Grammar" and "Structuring Rules Well" sections.

I considered that it would actually be intuitive and nice to have
"RHS" and "LHS" in the output, but "RHS" is disallowed in tree-sitter
because it's allowed to be empty, and I don't want to only have one of
them. Thus we remove "LHS" instead.